Use doubles instead of ints for the interval
svn path=/trunk/KDE/kdebase/workspace/; revision=864024

+++ gui/KSGAppletSettings.cc
@@ -104,12 +104,12 @@ void KSGAppletSettings::setSizeRatio( int value )
   mSizeRatio->setValue( value );
 }
 
-int KSGAppletSettings::updateInterval() const
+double KSGAppletSettings::updateInterval() const
 {
   return mInterval->value();
 }
 
-void KSGAppletSettings::setUpdateInterval( int value )
+void KSGAppletSettings::setUpdateInterval( double value )
 {
   mInterval->setValue( value );
 }
